third commit
In third commit I have made changes in navbar, where I align all the elements in the nav horizontally center. Added a bootstrap slider but commented it for later use. Made sections visually better using gird and flex, and added dummy content in it. Lastly styled the footer and the text inside it.